Changed FrameDelay to an int32 which is allowed by blueprints.  C# UHT wasn't generating an error.
#rb Chris.Kulla
#preflight 6310f54d31667117dfca0e28

[CL 21744521 by Tim Smith in ue5-main branch]

+++ Engine/Source/Developer/FunctionalTesting/Private/ScreenshotFunctionalTestBase.cpp
@@ -54,7 +54,7 @@ bool AScreenshotFunctionalTestBase::IsReady_Implementation()
 {
 	if ((GetWorld()->GetTimeSeconds() - RunTime) > ScreenshotOptions.Delay)
 	{
-		return (GFrameNumber - RunFrame) > ScreenshotOptions.FrameDelay;
+		return int32(GFrameNumber - RunFrame) > ScreenshotOptions.FrameDelay;
 	}
 
 	return false;
